#2355 create separate component for buttom GenSet
