+++ packages/ketcher-react/src/script/ui/views/modal/components/ExtendedTable/components/GenericGroups/GenSet.module.less
@@ -68,34 +68,3 @@
   column-gap: 4px;
 }
 
-.button {
-  width: 100%;
-  height: 100%;
-  font-size: @font-size-header;
-  line-height: @line-height-header;
-  vertical-align: middle;
-  padding: 0;
-  background: @color-background-primary;
-  border: 1.5px solid @color-grey-4;
-  border-radius: 4px;
-  color: fade(@color-text-secondary, 70%);
-
-  &:hover {
-    color: @color-text-secondary;
-    border: 1.5px solid @color-primary-green;
-  }
-
-  &:focus,
-  &.selected {
-    background-color: @color-primary-green;
-    color: @color-text-light;
-    border: 1.5px solid @color-primary-green;
-  }
-
-  &:disabled {
-    color: @color-grey-5;
-    border: 1.5px solid @color-grey-5;
-  }
-}
-
-
