+++ packages/ketcher-react/src/script/ui/views/modal/components/ExtendedTable/components/GenericGroups/components/ButtonGenSet/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ButtonGenSet'
